Move `CounterSpy` to its own file and group into a share scope
