+++ QuizMobileApp/QuizMobileAppTests/Game Engine/RestartGameUseCaseTests.swift
@@ -47,43 +47,5 @@ class RestartGameUseCaseTests: XCTestCase {
         trackForMemoryLeak(counter, file: file, line: line)
         
         return (sut, counter)
-    }
-    
-    private class CounterSpy: Counter {
-        var messages = [CounterResult]()
-        var seconds = 0
-        var startCompletions = [(CounterResult) -> Void]()
-        
-        init(seconds: Int) {
-            self.seconds = seconds
-        }
-        
-        func start(completion: @escaping (CounterResult) -> Void) {
-            guard seconds > 0 else { return }
-            messages.append(.start)
-            
-            while seconds > 0 {
-                messages.append(.currentSecond(seconds))
-                startCompletions.append(completion)
-                
-                seconds -= 1
-            }
-        }
-        
-        func startGameMessage(at index: Int = 0) {
-            guard startCompletions.count > 0 else { return }
-            
-            startCompletions[index](.start)
-        }
-        
-        func sendCurrentSecond(at index: Int = 0) {
-            guard startCompletions.count > 0 else { return }
-            
-            startCompletions[index](.currentSecond(seconds))
-        }
-        
-        func reset() {
-            messages.append(.reset)
-        }
-    }
+    }        
 }
